+++ packages/@expo/metro-config/src/serializer/treeShakeSerializerPlugin.ts
@@ -78,54 +78,64 @@ function isEmptyModule(value: Module<MixedOutput>): boolean {
   });
 }
 
-// Collect a list of exports that are not used within the module.
+// Collect a map of unused identifiers to their exported names
+// Identifiers are only included in the map if they are exported
 function getExportsThatAreNotUsedInModule(ast: types.File) {
-  const exportedIdentifiers = new Set<string>();
+  const exportedIdentifiers = new Map<string, string[]>();
   const usedIdentifiers = new Set<string>();
-  const unusedExports: string[] = [];
+  const unusedExports = new Map<string, string[]>();
 
-  // First pass: collect all export identifiers
+  const recordIdentifier = (local: string, exported: string) => {
+    exportedIdentifiers.set(local, (exportedIdentifiers.get(local) ?? []).concat(exported));
+  };
+
   traverse(ast, {
-    ExportNamedDeclaration(path) {
-      const { declaration, specifiers } = path.node;
-      if (declaration) {
-        if ('declarations' in declaration && declaration.declarations) {
+    Declaration(path) {
+      const declaration = path.node;
+      if (types.isVariableDeclaration(declaration)) {
+        if (types.isExportNamedDeclaration(path.parent)) {
           declaration.declarations.forEach((decl) => {
             if (types.isIdentifier(decl.id)) {
-              exportedIdentifiers.add(decl.id.name);
+              recordIdentifier(decl.id.name, decl.id.name);
             }
           });
-        } else if ('id' in declaration && types.isIdentifier(declaration.id)) {
-          exportedIdentifiers.add(declaration.id.name);
+        }
+      } else if ('id' in declaration && types.isIdentifier(declaration.id)) {
+        if (types.isExportNamedDeclaration(path.parent)) {
+          recordIdentifier(declaration.id.name, declaration.id.name);
+        } else if (types.isExportDefaultDeclaration(path.parent)) {
+          recordIdentifier(declaration.id.name, 'default');
         }
       }
-      specifiers.forEach((spec) => {
-        if (types.isIdentifier(spec.exported)) {
-          exportedIdentifiers.add(spec.exported.name);
-        }
-      });
     },
-
-    ExportDefaultDeclaration(path) {
-      // Default exports need to be handled separately
-      // Assuming the default export is a function or class declaration:
-      if ('id' in path.node.declaration && types.isIdentifier(path.node.declaration.id)) {
-        exportedIdentifiers.add(path.node.declaration.id.name);
-      }
+    ExportSpecifier(path) {
+      const exportedIdentifier = types.isIdentifier(path.node.exported)
+        ? path.node.exported.name
+        : path.node.exported.value;
+      recordIdentifier(path.node.local.name, exportedIdentifier);
     },
-  });
-
-  // Second pass: find all used identifiers
-  traverse(ast, {
     ReferencedIdentifier(path) {
-      usedIdentifiers.add(path.node.name);
+      if (types.isExportSpecifier(path.parent)) {
+        const exportedIdentifier = types.isIdentifier(path.parent.exported)
+          ? path.parent.exported.name
+          : path.parent.exported.value;
+        recordIdentifier(path.parent.local.name, exportedIdentifier);
+      } else if (types.isExportDefaultSpecifier(path.parent)) {
+        recordIdentifier(path.node.name, 'default');
+      } else {
+        const binding = path.scope.getBinding(path.node.name);
+        const isModuleScope = binding && !binding.scope.parent;
+        if (isModuleScope) {
+          usedIdentifiers.add(path.node.name);
+        }
+      }
     },
   });
 
   // Determine which exports are unused
-  exportedIdentifiers.forEach((exported) => {
-    if (!usedIdentifiers.has(exported)) {
-      unusedExports.push(exported);
+  exportedIdentifiers.forEach((exported, local) => {
+    if (!usedIdentifiers.has(local)) {
+      unusedExports.set(local, exported);
     }
   });
 
@@ -641,110 +651,76 @@ export async function treeShakeSerializer(
       // Collect a list of exports that are not used within the module.
       const possibleUnusedExports = getExportsThatAreNotUsedInModule(ast);
 
-      // Traverse exports and mark them as used or unused based on if inverse dependencies are importing them.
+      // Remove declarations and specifiers referenced by exports if no inverse dependencies are importing them.
+      // While entering, removes variable declarators and export specifiers.
+      // While exiting, removes empty export named declarations and variable declarations.
       traverse(ast, {
-        ExportDefaultDeclaration(path) {
-          if (possibleUnusedExports.includes('default') && !isExportUsed('default')) {
-            // TODO: Update source maps
-            markUnused(path);
-          }
-        },
-
-        ExportNamedDeclaration(path) {
-          const importModuleId = path.node.source?.value;
-
-          // Remove specifiers, e.g. `export { Foo, Bar as Bax }`
-          if (types.isExportNamedDeclaration(path.node)) {
-            for (let i = 0; i < path.node.specifiers.length; i++) {
-              const specifier = path.node.specifiers[i];
-              if (
-                types.isExportSpecifier(specifier) &&
-                types.isIdentifier(specifier.local) &&
-                types.isIdentifier(specifier.exported) &&
-                possibleUnusedExports.includes(specifier.exported.name) &&
-                !isExportUsed(specifier.exported.name)
-              ) {
-                // Remove specifier
-                path.node.specifiers.splice(i, 1);
-
-                needsImportReindex = true;
-                i--;
-              }
-            }
-          }
-
-          // Remove the entire node if the export has been completely removed.
-
-          const declaration = path.node.declaration;
-
-          if (types.isVariableDeclaration(declaration)) {
-            declaration.declarations = declaration.declarations.filter((decl) => {
-              if (decl.id.type === 'Identifier') {
-                if (possibleUnusedExports.includes(decl.id.name) && !isExportUsed(decl.id.name)) {
+        Declaration: {
+          enter(path) {
+            const declaration = path.node;
+            if ('id' in declaration && types.isIdentifier(declaration.id)) {
+              const binding = path.scope.getBinding(declaration.id.name);
+              const isModuleScope = binding && !binding.scope.parent;
+              if (isModuleScope) {
+                const exportedIdentifiers = possibleUnusedExports.get(declaration.id.name);
+                if (exportedIdentifiers && !exportedIdentifiers.some(isExportUsed)) {
                   // TODO: Update source maps
-                  debug(
-                    `mark remove (type: var, depth: ${depth}):`,
-                    decl.id.name,
-                    'from:',
-                    value.path
-                  );
-
-                  // Account for variables, and classes which may contain references to other exports.
+                  markUnused(path);
                   shouldRecurseUnusedExports = true;
-                  return false; // Remove this declaration
                 }
               }
-              return true; // Keep this declaration
-            });
-
-            // If all declarations were removed, remove the entire path
-            if (declaration.declarations.length === 0) {
-              markUnused(path);
             }
-          } else if (declaration && 'id' in declaration && types.isIdentifier(declaration.id)) {
-            // function, class, etc.
-            if (
-              possibleUnusedExports.includes(declaration.id.name) &&
-              !isExportUsed(declaration.id.name)
-            ) {
-              debug(
-                `mark remove (type: function, depth: ${depth}):`,
-                declaration.id.name,
-                'from:',
-                value.path
-              );
-              // TODO: Update source maps
-              markUnused(path);
+          },
+        },
 
-              // Code inside of an unused export may affect other exports in the module.
-              // e.g.
-              //
-              // export const a = () => {}
-              //
-              // // Removed `b` -> recurse for `a`
-              // export const b = () => { a() };
-              //
-              shouldRecurseUnusedExports = true;
-            }
-          }
-
-          // If export from import then check if we can remove the import.
-          if (importModuleId) {
-            if (path.node.specifiers.length === 0) {
-              const removeRequest = disconnectGraphNode(value, importModuleId, {
-                // Due to a quirk in how we've added tree shaking, export-all is expanded and could overlap with existing exports in the same module.
-                // If we find that the existing export has a default export then we should skip removing the node entirely.
-                bailOnDuplicateDefaultExport: true,
-              });
-              if (removeRequest.removed) {
-                dirtyImports.push(removeRequest.path);
+        ExportNamedDeclaration: {
+          exit(path) {
+            if (!path.node.declaration && path.node.specifiers.length === 0) {
+              // If export from import then check if we can remove the import.
+              const importModuleId = path.node.source?.value;
+              if (importModuleId) {
+                const removeRequest = disconnectGraphNode(value, importModuleId, {
+                  // Due to a quirk in how we've added tree shaking, export-all is expanded and could overlap with existing exports in the same module.
+                  // If we find that the existing export has a default export then we should skip removing the node entirely.
+                  bailOnDuplicateDefaultExport: true,
+                });
+                if (removeRequest.removed) {
+                  dirtyImports.push(removeRequest.path);
+                }
               }
+
               // TODO: Update source maps
-              // We still want to remove the empty `export {} from 'a'` declaration even if the graph node was kept
-              // due to the duplicate default export
               markUnused(path);
             }
-          }
+          },
+        },
+
+        ExportSpecifier: {
+          enter(path) {
+            const specifier = path.node;
+            const localIdentifier = specifier.local.name;
+            const exportedIdentifier = types.isIdentifier(specifier.exported)
+              ? specifier.exported.name
+              : specifier.exported.value;
+            if (possibleUnusedExports.has(localIdentifier) && !isExportUsed(exportedIdentifier)) {
+              path.remove();
+              needsImportReindex = true;
+            }
+          },
+        },
+
+        VariableDeclarator: {
+          enter(path) {
+            const isModuleScope = !path.scope.parent;
+            if (isModuleScope && types.isIdentifier(path.node.id)) {
+              const exportedIdentifiers = possibleUnusedExports.get(path.node.id.name);
+              if (exportedIdentifiers && !exportedIdentifiers.some(isExportUsed)) {
+                // TODO: Update source maps
+                markUnused(path);
+                shouldRecurseUnusedExports = true;
+              }
+            }
+          },
         },
       });
     }